+++ src/org/soulspace/overarch/domain/views/component_view.clj
@@ -19,7 +19,7 @@
          (element->boundary (:el e))           ; has a boundary mapping for this diagram-type
          (or (el/internal? e)                  ; is internal
              (view/expand-external-spec view)) ; should be expanded
-         (not= true (:collapsed e))))))
+         (not (el/collapsed? e))))))
 
 (defn render-model-node?
   "Returns true if the `model` node `e` is rendered in the component `view`."